tests/testlib/_crypto.py
# SPDX-FileCopyrightText: 2025 Simon Holesch <simon@holesch.de>
#
# SPDX-License-Identifier: AGPL-3.0-or-later

# ruff: noqa: F401
from Cryptodome.Cipher import ChaCha20_Poly1305 as XChaCha20Poly1305
from cryptography.hazmat.primitives.asymmetric.x25519 import (
    X25519PrivateKey,
    X25519PublicKey,
)
from cryptography.hazmat.primitives.ciphers.aead import ChaCha20Poly1305
from cryptography.hazmat.primitives.hashes import BLAKE2s
from cryptography.hazmat.primitives.kdf.hkdf import HKDF
