requirements.txt
# SPDX-FileCopyrightText: 2025 Simon Holesch <simon@holesch.de>
#
# SPDX-License-Identifier: AGPL-3.0-or-later

# test dependencies
pytest
cryptography
pycryptodomex
scapy

# static analysis
black
isort
ruff
